add project comments (#185)
Show project comments on the dashboard. The projects are hardcoded, but once you have selected a project and scroll down to comments, it will show the comments for that project ID (comes from API)

Also, I have set a limit to the comment length to prevent people from putting in essays!