Updating contributing guidelines
I have now simplified the labels list.   As discussed,  I have updated the contributing guidelines to reflect the changes and explain the different ways in which people can contribute.

The plan is to keep this page up to date with any changes/updates as we continue the collaboration with the collective.