Refactor footer path matching to use matchRoutes instead of matchPath
Uses matchRoutes instead of useMatch to compare footer paths and decide whether to display the footer. Unlike useMatch, matchRoutes allows us to compare a given location to an array of route objects without looping through each match. matchRoutes is suited for numerous routes and useMatch for one-to-one route matching. We have been using useMatch because @reach/router does not have support for matchRoutes.