Fix boolean parsing of SMTP configuration from environment variables. (#5861)
-------------------------------------------------
In the previous implementation, there was an issue with parsing the boolean values of the SMTP configuration from environment variables. The use of bool() on the string returned by os.getenv() caused both MAIL_USE_TLS and MAIL_USE_SSL to be set to True, regardless of the actual environment variable values.