Check if project exists on task action endpoints so that we can return appropriate error messages
In the prev implementation even if the project id is non existent  error message was returned as project_id and task_id are primary keys for tasks. THis addition will check if the project exists before checking for task so we would already know if project exists and return  error.