Update requires-python version to ">=3.10,<=3.11"
This commit modifies the requires-python constraint in the pyproject.toml file to ensure that the Python version used for this project remains between 3.10 and 3.11. By locking the Python version in this range, we guarantee that the generated lock file will be based on Python 3.10, which aligns with the production setup requirements. This change helps maintain consistency and stability across environments and ensures compatibility with the selected Python version.